move export.ts
