+++ packages/studio-base/src/panels/ThreeDeeRender/export.ts
